Here is a copy of the COMPLETE directory found in /export/swift3
This >should be< the working, latest version of the code, but it is hard to tell since everyone stopped using revision control at some point.

Mostly just a copy of the COMPLETE files ( *.c *.h makefiles) over the repos files.

Modifications include fixing the makefile in jsd for hardcoded paths.

On the bright side it compiles now!

--
Christian